+++ core/focus/messages.go
@@ -1,0 +1,6 @@
+package focus
+
+type FocusChangedMsg struct {
+	From int
+	To   int
+}